+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `Popup` to appear when the `MiniCart` button is clicked instead of on hover.
+
 ### Fixed 
 - `maxQuantity` undefined warning.
 - Loading mixed (insecure) display content warning.